npx example works
